requirements.txt
garminconnect>=0.2.20
garth>=0.4.46
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
